Split it Delivery and Client
